split ui classes into their own files, kivy apparently doesn't care about namespaces
